add update params related types
